% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferencePanel.R
\name{buildReferencePanel}
\alias{buildReferencePanel}
\title{Function to build a reference panel.
Celltype and replicate information must be included in the colnames of the bulk.rna.data.object, separted by an underscore ("_"), e.g.: Bnaive_R1.}
\usage{
buildReferencePanel(
  bulk.rna.data,
  fc.thrs.general = 6,
  fc.thrs.specific = 2,
  fdr.thrs = 0.01,
  cut_height = 1.1,
  gene.nomenclature = "SYMBOL",
  species = "HUMAN",
  filename = "my_reference_panel.rds",
  verbose = FALSE
)
}
\arguments{
\item{bulk.rna.data}{Bulk RNA sequencing data}

\item{fc.thrs.general}{absolulte log fold change threshold applied for coarse marker detection (default 6)}

\item{fc.thrs.specific}{absolulte log fold change threshold applied for detailed marker detection (default 2)}

\item{fdr.thrs}{FDR threshold (default 0.01)}

\item{cut_height}{to cut the clustering of the reference data sets (default 1.1)}

\item{gene.nomenclature}{Type of gene nomenclature. "ENS" for Ensembl IDs, "SYMBOL" for gene symbols. Default is "SYMBOL".}

\item{species}{"HUMAN" for human data, "MOUSE" for mouse data. Default is "HUMAN".}

\item{filename}{file name of reference panel (should end with .rds)}

\item{verbose}{Generate debug output and figures (default FALSE)}
}
\value{
reference panel as a data.frame
}
\description{
Function to build a reference panel.
Celltype and replicate information must be included in the colnames of the bulk.rna.data.object, separted by an underscore ("_"), e.g.: Bnaive_R1.
}
