% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cellClust.R
\name{cellClust}
\alias{cellClust}
\title{Generate cell clusters}
\usage{
cellClust(obj_in, method = "hclust", deepSplit_wgcna = 1,
  min_group_Size_wgcna = 5)
}
\arguments{
\item{obj_in}{data object.}

\item{method}{can only be "hclust" (default).}

\item{deepSplit_wgcna}{integer value indicating how deep the dendrogram should be cut. Values can be 0, 1 (default), 2, 3 and 4.}

\item{min_group_Size_wgcna}{integer value indicating the minimum size of the resulting clusters. Default is 5.}
}
\value{
data object.
}
\description{
Hierarchical clustering and dynamic tree cutting.
}
\examples{
data_obj = cellClust(data_obj);
}

