% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cellNormalize.R
\name{cellNormalize}
\alias{cellNormalize}
\title{Normalize data}
\usage{
cellNormalize(obj_in, method = "no_norm")
}
\arguments{
\item{obj_in}{data object.}

\item{method}{can choose between "no_norm" (Default, no normalization would be performed) and "scQ" (an adjusted quantile normalization for single cell data).}
}
\value{
data object.
}
\description{
Normalize single cell RNAseq data.
}
\examples{
data_obj = cellNormalize(data_obj);
}

