% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePairDE.R
\name{ComputePairWiseDE}
\alias{ComputePairWiseDE}
\title{Find markers (differentially expressed genes) between two group of cells.}
\usage{
ComputePairWiseDE(
  object,
  cells.1 = NULL,
  cells.2 = NULL,
  features = NULL,
  logfc.threshold = 1.5,
  test.use = "wilcox",
  min.pct = 0.25,
  min.diff.pct = -Inf,
  verbose = TRUE,
  only.pos = FALSE,
  max.cells.per.ident = Inf,
  random.seed = 1,
  min.cells.group = 3,
  pseudocount.use = 1,
  MeanExprsThrs = 0,
  p.adjust.methods = "BH"
)
}
\arguments{
\item{object}{dataMatrix of genes (rows) x cells (columns) expression matrix (log normalized value)}

\item{cells.1}{Vector of cell names belonging to group 1}

\item{cells.2}{Vector of cell names belonging to group 2}

\item{features}{Genes to test. Default is NULL which mean to use all genes}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is log(1.5)}

\item{test.use}{Denotes which test to use. Available options are:
\itemize{
 \item{"wilcox"} : Identifies differentially expressed genes between two
 groups of cells using a Wilcoxon Rank Sum test (default)
 \item{"bimod"} : Likelihood-ratio test for single cell gene expression,
 (McDavid et al., Bioinformatics, 2013)
 \item{"roc"} : Identifies 'markers' of gene expression using ROC analysis.
 For each gene, evaluates (using AUC) a classifier built on that gene alone,
 to classify between two groups of cells. An AUC value of 1 means that
 expression values for this gene alone can perfectly classify the two
 groupings (i.e. Each of the cells in cells.1 exhibit a higher level than
 each of the cells in cells.2). An AUC value of 0 also means there is perfect
 classification, but in the other direction. A value of 0.5 implies that
 the gene has no predictive power to classify the two groups. Returns a
 'predictive power' (abs(AUC-0.5) * 2) ranked matrix of putative differentially
 expressed genes.
 \item{"t"} : Identify differentially expressed genes between two groups of
 cells using the Student's t-test.
 }}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.25}

\item{min.diff.pct}{only test genes that show a minimum difference in the
fraction of detection between the two groups. Set to -Inf by default}

\item{verbose}{Print a progress bar once expression testing begins}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{max.cells.per.ident}{Down sample each identity class to a max number.
Default is no downsampling. Not activated by default (set to Inf)}

\item{random.seed}{Random seed for downsampling. default is 1}

\item{min.cells.group}{Minimum number of cells in one of the groups}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when calculating logFC. 1 by default.}

\item{MeanExprsThrs}{a minimum expression threshold of average cluster expression for a gene to be considered a DE gene.
the mean expression value is in the linear scale!}

\item{p.adjust.methods}{correction method for calculating qvalue. default is BH (or FDR)}
}
\value{
returnObj
}
\description{
Find markers (differentially expressed genes) between two group of cells.
}
\author{
Ignasius Joanito (Modified from Seurat FindMarkers)
}
