% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUMAP.R
\name{computeUMAP}
\alias{computeUMAP}
\title{Compute a umap projection into 2 or 3 dimensions}
\usage{
computeUMAP(rca.obj, nDIMS = 2)
}
\arguments{
\item{rca.obj}{RCA object}

\item{nDIMS}{number of UMAP dimensions}
}
\description{
Compute a umap projection into 2 or 3 dimensions
}
