% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corrSim}
\alias{corrSim}
\title{Simulate example data}
\usage{
corrSim(corr, N = 1e+06, seed = 123)
}
\arguments{
\item{An}{11x11 correlation matrix}
}
\description{
Simulate MVN example data from the \code{corr} data, which are a correlation
matrix
}
