% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRCAObject.R
\name{createRCAObject}
\alias{createRCAObject}
\title{Create RCA object}
\usage{
createRCAObject(rawData, dataIsNormalized = FALSE)
}
\arguments{
\item{rawData}{A matrix of expression values}

\item{dataIsNormalized}{A flag indicating whether the provided data is already normalized}
}
\value{
RCA object.
}
\description{
Create RCA object
}
