% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRCAObjectFrom10X.R
\name{createRCAObjectFrom10X}
\alias{createRCAObjectFrom10X}
\title{Create RCA Object from 10X data}
\usage{
createRCAObjectFrom10X(dataDir, min.barcode.umi = 100)
}
\arguments{
\item{dataDir}{Directory containing 10X data}

\item{min.barcode.umi}{Minimum UMIs needed for a barcode to be considered. Default is 100.}
}
\value{
RCA object.
}
\description{
Create RCA Object from 10X data
}
