% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClust.R
\name{dataClust}
\alias{dataClust}
\title{Generate cell clusters using hierarchical clustering and dynamic tree cutting.}
\usage{
dataClust(rca.obj, deepSplitValues = 1, minClustSize = 5)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{deepSplitValues}{integer vector indicating how deep the dendrogram should be cut. Values can range from 0 to 4.}

\item{minClustSize}{integer value indicating the minimum size of the resulting clusters. Default is 5.}
}
\value{
RCA object.
}
\description{
Generate cell clusters using hierarchical clustering and dynamic tree cutting.
}
