% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataConstruct.R
\name{dataConstruct}
\alias{dataConstruct}
\title{Construct data object}
\usage{
dataConstruct(input_data)
}
\arguments{
\item{input_data}{A data frame of expression values (FPKM, TPM, UMI counts ...), with rows representing genes and columns representing cells. Note the current version of RCA only accepts gene names in the following format: "GenomeLocation_HGNCGeneName_EnsembleID", from which the "HGNCGeneName" is extracted for RCA analysis. For input data with only HGNC names, the users need to attach two strings to the HGNC names to make them into the "XXXX_HGNCGeneNames_ENSYYYY" format.}
}
\value{
The data object for the following RCA analysis.
}
\description{
Construct a data object for the analysis
}
\examples{
input_data = read.csv("fpkm_data.csv",row.names=1);
data_obj = dataConstruct(input_data);
}

