% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePairDE.R
\name{dataDE}
\alias{dataDE}
\title{Compute pairwise DE genes for supervised clustering result.}
\usage{
dataDE(
  rca.obj,
  logFoldChange = 1.5,
  method = "wilcox",
  mean.Exp = 0.5,
  deepsplit = 1,
  min.pct = 0.25,
  min.diff.pct = -Inf,
  random.seed = 1,
  min.cells.group = 3,
  pseudocount.use = 1,
  p.adjust.methods = "BH",
  top.genes.per.cluster = 10,
  pairwise = FALSE,
  nCores = 1
)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{logFoldChange}{Log fold change required to call gene DE (default is log(1.5)).}

\item{method}{Denotes which test to use. Available options are:
\itemize{
 \item{"wilcox"} : Identifies differentially expressed genes between two
 groups of cells using a Wilcoxon Rank Sum test (default)
 \item{"bimod"} : Likelihood-ratio test for single cell gene expression,
 (McDavid et al., Bioinformatics, 2013)
 \item{"roc"} : Identifies 'markers' of gene expression using ROC analysis.
 For each gene, evaluates (using AUC) a classifier built on that gene alone,
 to classify between two groups of cells. An AUC value of 1 means that
 expression values for this gene alone can perfectly classify the two
 groupings (i.e. Each of the cells in cells.1 exhibit a higher level than
 each of the cells in cells.2). An AUC value of 0 also means there is perfect
 classification, but in the other direction. A value of 0.5 implies that
 the gene has no predictive power to classify the two groups. Returns a
 'predictive power' (abs(AUC-0.5) * 2) ranked matrix of putative differentially
 expressed genes.
 \item{"t"} : Identify differentially expressed genes between two groups of
 cells using the Student's t-test.
 }}

\item{mean.Exp}{Minimum mean expression of a gene to be considered in the DE gene calculation (default 0.5))}

\item{deepsplit}{If hclust was used for clustering, the desired deepsplit can be specified here. Values can range from 0 to 4 (default is 1).}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed (default is 0.25).}

\item{min.diff.pct}{only test genes that show a minimum difference in the
fraction of detection between the two groups (-Inf by default).}

\item{random.seed}{Random seed for downsampling (default 1).}

\item{min.cells.group}{Minimum number of cells in one of the groups (default 3).}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when calculating logFC (default 1).}

\item{p.adjust.methods}{correction method for calculating q-value (default is BH ).}

\item{top.genes.per.cluster}{Number of top DE genes to be considered per cluster(default 10).}

\item{pairwise}{Flag indicating whether DE genes should be compared derived in pairwise manner or 1 cluster vs all others (default).}

\item{nCores}{Number of cores to used for parallel computation (default 1).}
}
\value{
RCA object.
}
\description{
Compute pairwise DE genes for supervised clustering result.
}
\examples{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
RCA.pbmcs <- dataClust(RCA.pbmcs)
RCA.pbmcs <- dataDE(RCA.pbmcs)

}
