\name{dataFramePalette}
\alias{dataFramePalette}
\title{
Quantize RGB colours into a palette.
}
\description{
  This function uses k-means clustering to extract a palette from a
  data.frame representing the RGB values of pixels in an image. This is
  not the most state-of-the-art method of extracting a palette, but
  works as well as most photo editing software.
}
\usage{
dataFramePalette(dataframe, count)
}
\arguments{
  \item{dataframe}{
    A data.frame representing the colours of the pixels in the image,
    generated by \code{imageToDataFrame()}.
  }
  \item{count}{
    The number of colours to try to extract
  }
}
\value{
  A vector of colorspace library RGB objects.
}
\examples{
library(colorspace)
imgdir<-paste(system.file(package = "CulturalAnalytics"), "images", sep = "/")
imgs<-paste(imgdir, dir(path = imgdir, pattern = ".jpg"), sep = "/")
img<-readImage(imgs[1])
df<-imageToDataFrame(img)
palette<-dataFramePalette(df, 8)
}
