% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLogNormalise.R
\name{dataLogNormalise}
\alias{dataLogNormalise}
\title{Log-transform and normalise data by sequencing depth}
\usage{
dataLogNormalise(rca.obj, scale.factor = 10000)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{scale.factor}{scaling factor for log-normalization}
}
\value{
RCA object.
}
\description{
Log-transform and normalise data by sequencing depth
}
