% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProject.R
\name{dataProject}
\alias{dataProject}
\title{Compute Reference Component features for clustering analysis}
\usage{
dataProject(
  rca.obj,
  method = "GlobalPanel",
  customPath = NULL,
  corMeth = "pearson",
  power = 4,
  scale = T,
  min.cell.number.expressing = 1
)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{method}{Either "GlobalPanel" (default) or "GlobalPanel_CellTypes", "ColonEpiPanel","MonacoPanel","ENCODEMousePanel","ENCODEHumanPanel","ZhangMouseBrainPanel","NovershternPanel","NovershternTCellPanel","MonacoTCellPanel","MonacoBCellPanel","MonacoMonoPanel", "CITESeqPanel" or "Custom"}

\item{customPath}{directory path (including filename) to any custom panel stored in RDS format. Only used if method == "Custom".}

\item{corMeth}{Any of the correlation measures supported by R (default pearson)}

\item{power}{power to raise up to for the RCA features before clustering (default 4)}

\item{scale}{True if the data should be scaled, False otherwise (default TRUE)}

\item{min.cell.number.expressing}{Minimum number of cells (0-100) expressing a gene such that it is considered in the projection step (default 1).}
}
\value{
RCA object.
}
\description{
Compute Reference Component features for clustering analysis
}
\examples{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
print(RCA.pbmcs)

}
