% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProject.R
\name{dataProjectMultiPanel}
\alias{dataProjectMultiPanel}
\title{Compute Reference Component features for clustering analysis}
\usage{
dataProjectMultiPanel(
  rca.obj,
  method = list("GlobalPanel"),
  customPath = NULL,
  corMeth = "pearson",
  power = 4,
  scale = T
)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{method}{List of panel identifiers containing "GlobalPanel"(default), "ColonEpiPanel", "MonacoPanel","ENCODEMousePanel","ENCODEHumanPanel","ZhangMouseBrainPanel","NovershternPanel"}

\item{customPath}{list of paths (including filename) to any custom panel stored in RDS format. Only used if method == "Custom".}

\item{corMeth}{Any of the correlation measures supported by R, defaults to pearson}

\item{power}{power to raise up to for the RCA features before clustering, default is 4}
}
\value{
RCA object.
}
\description{
Compute Reference Component features for clustering analysis
}
