% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSNN.R
\name{dataSNN}
\alias{dataSNN}
\title{Generate cell clusters using hierarchical clustering and dynamic tree cutting.}
\usage{
dataSNN(rca.obj, k = 50, eps = 20, minPts = 10)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{deepSplitValues}{integer vector indicating how deep the dendrogram should be cut. Values can range from 0 to 4.}

\item{minClustSize}{integer value indicating the minimum size of the resulting clusters. Default is 5.}
}
\value{
RCA object.
}
\description{
Generate cell clusters using hierarchical clustering and dynamic tree cutting.
}
