% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{dataTransform}
\alias{dataTransform}
\title{Data Transformation}
\usage{
dataTransform(obj_in, method = "log10")
}
\arguments{
\item{obj_in}{data object.}

\item{method}{can only be "log10" (default), indicating a log10 transformation of the data.}
}
\value{
data object.
}
\description{
Transform raw values into log10 scale
}
\examples{
data_obj = dataTransform(data_obj);
}

