% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSClust.R
\name{elbowPlot}
\alias{elbowPlot}
\title{Generate elbow plot for a PCA of the projection matrix.}
\usage{
elbowPlot(rca.obj, nPCs = 50, filename = "Projection_Elbow.png", approx = F)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{nPCs}{Number of PCs to be used if distance should be computed in PC embedding of the projection (default 0, computation of distance in projection space)}

\item{approx}{Flag indicating wheter an approximation of the SVD should be used}
}
\value{
NULL.
}
\description{
Generate elbow plot for a PCA of the projection matrix.
}
