% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellTypeFromProjection.R
\name{estimateCellTypeFromProjection}
\alias{estimateCellTypeFromProjection}
\title{Estimate the most likely cell type from the projection to the reference panel}
\usage{
estimateCellTypeFromProjection(rca.obj, confidence = NULL)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{confidence:}{a parameter indicating the difference between z-scores. If the difference is below this threshold, the cell type will be set to unknown. Default is NULL.}
}
\value{
RCA object.
}
\description{
Estimate the most likely cell type from the projection to the reference panel
}
