% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/featureConstruct.R
\name{featureConstruct}
\alias{featureConstruct}
\title{Generate Reference Component featrues}
\usage{
featureConstruct(obj_in, method = "GlobalPanel", power = 4)
}
\arguments{
\item{obj_in}{data object.}

\item{method}{choose from three RCA options: "GlobalPanel"(default), "ColonEpitheliumPanel", or "SelfProjection".}

\item{power}{power to raise up to for the RCA features before clustering. Default is 4.}
}
\value{
data object
}
\description{
Compute Reference Component features for clustering analysis
}
\examples{
data_obj = featureConstruct(data_obj);
}

