% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneFilt.R
\name{geneFilt}
\alias{geneFilt}
\title{Filt genes}
\usage{
geneFilt(obj_in, method = "default")
}
\arguments{
\item{obj_in}{data object}
}
\value{
data object
}
\description{
remove genes with too low expression levels in most cells.
}
\examples{
data_obj = geneFilt(data_obj);
}

