% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterSpaceSeurat.R
\name{parameterSpaceSeurat}
\alias{parameterSpaceSeurat}
\title{Generates a data frame showcasing the dependence between resolution and cluster quantity}
\usage{
parameterSpaceSeurat(rca.obj, stepsize = 0.1, filename = NULL)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{stepsize}{Stepsize used to generate an overview on clustering results (default 0.1)}

\item{filename}{name of the pdf file produced that visuzalizes the parameter space (default NULL)}
}
\value{
list of a a data frame holding parameter values and resulting cluster numbers and the parameter figure
}
\description{
Generates a data frame showcasing the dependence between resolution and cluster quantity
}
\examples{
\dontrun{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
parameterSpaceSeurat(RCA.pbmcs)
}

}
