% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterSpaceSNN.R
\name{parameterSpaceSNN}
\alias{parameterSpaceSNN}
\title{Generates an overview on the dependence of SNN parameters to the number of clusters}
\usage{
parameterSpaceSNN(
  rca.obj,
  kL = base::c(30:50),
  epsL = base::c(5:20),
  minPtsL = base::c(5:10),
  filename = NULL
)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{kL}{vector of k values to be used (default is 30 to 50).}

\item{epsL}{vector of values indicating the number of neighbors between two nodes such that cells are shared (default is 5 to 20).}

\item{minPtsL}{Vector of minimum points with at least eps neighbors to be considered a core point (default is 5 to 10)}

\item{filename}{name of the html file produced that visualizes the parameter space (default NULL)}
}
\value{
list of a data frame holding parameter values and resulting cluster numbers and the 3D parameter figure
}
\description{
Generates an overview on the dependence of SNN parameters to the number of clusters
}
\examples{
\dontrun{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
parameterSpaceSNN(RCA.pbmcs)
}
}
