% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_small_counts-data.R
\docType{data}
\name{pbmc_small_counts}
\alias{pbmc_small_counts}
\title{Small PBMC dataset}
\format{
An object of class \code{dgCMatrix} with 230 rows and 80 columns.
}
\usage{
pbmc_small_counts
}
\description{
Raw UMI counts from the pbmc_small object of the Seurat package.
}
\references{
Hao et al. (2020). bioRxiv (doi: 10.1101/2020.10.12.335331).
}
\keyword{datasets}
