% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performClusterSpecificQC.R
\name{performClusterSpecificQC}
\alias{performClusterSpecificQC}
\title{Filter the dataset and remove genes and cells that could be of poor quality using user-defined thresholds}
\usage{
performClusterSpecificQC(
  rca.obj,
  cluster.labels,
  nGene.low.thresholds = NULL,
  nGene.high.thresholds = NULL,
  nUMI.low.thresholds = NULL,
  nUMI.high.thresholds = NULL,
  pMito.low.thresholds = NULL,
  pMito.high.thresholds = NULL
)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{cluster.labels}{vector of cluster labels}

\item{nGene.low.thresholds}{numeric vector with lower nGene thresholds. Length of vector should be length(unique(cluster.labels)).}

\item{nGene.high.thresholds}{numeric vector with upper nGene thresholds. Length of vector should be length(unique(cluster.labels)).}

\item{nUMI.low.thresholds}{numeric vector with lower nUMI thresholds. Length of vector should be length(unique(cluster.labels)).}

\item{nUMI.high.thresholds}{numeric vector withupper nUMI thresholds. Length of vector should be length(unique(cluster.labels)).}

\item{pMito.low.thresholds}{numeric vector with lower pMito thresholds. Length of vector should be length(unique(cluster.labels)).}

\item{pMito.high.thresholds}{numeric vector and upper pMito thresholds. Length of vector should be length(unique(cluster.labels)).}
}
\value{
RCA object.
}
\description{
Filter the dataset and remove genes and cells that could be of poor quality using user-defined thresholds
}
