% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterQuality.R
\name{plotClusterQuality}
\alias{plotClusterQuality}
\title{Perform cluster specific QC}
\usage{
plotClusterQuality(
  rca.obj,
  cluster.labels,
  width = 20,
  height = 20,
  filename = "RCA_Cluster_Quality.pdf"
)
}
\arguments{
\item{rca.obj}{RCA object}

\item{cluster.labels}{vector of cluster labels for each cell}

\item{width}{width of plot in inches (default 20).}

\item{height}{height of plot in inches (default 20).}

\item{filename}{file name of saved scatter plots}
}
\description{
Perform cluster specific QC
}
\examples{
\dontrun{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
RCA.pbmcs <- dataClust(RCA.pbmcs)
plotClusterQuality(RCA.pbmcs,
RCA.pbmcs$clustering.out$dynamicColorsList[[1]])
}
}
