\name{plotColourCloud}
\alias{plotColourCloud}
\title{
  Plot a colour histogram as a colour cloud.
}
\description{
  Function to plot a ``colour cloud'' spatial representation of an RGB
  colour histogram of an image.
}
\usage{
plotColourCloud(hist, title = "", radius = 0.005, dotSize = 1)
}
\arguments{
  \item{hist}{
    A colour histogram generated by \code{rgbImageHistogram()} .
  }
  \item{title}{
    The main title of the plot.
  }
  \item{radius}{
    The jitter radius of dots in the cloud.
  }
  \item{dotSize}{
    The size of each dot in the colour cloud.
  }
}
\value{
  None
}
\references{
  See: http://cns.bu.edu/~gsc/ColorHistograms.html
}

\examples{
imgdir<-paste(system.file(package = "CulturalAnalytics"), "images", sep = "/")
imgs<-paste(imgdir, dir(path = imgdir, pattern = ".jpg"), sep = "/")
img<-readImage(imgs[1])
rgbdf<-imageToDataFrame(img)
rgbhist<-rgbImageHist(rgbdf)
plotColourCloud(rgbhist)
}