% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEHeatmap.R
\name{plotDEHeatmap}
\alias{plotDEHeatmap}
\title{Plot heatmap of DE genes}
\usage{
plotDEHeatmap(
  rca.obj,
  scale = FALSE,
  width = 20,
  height = 20,
  filename = "RCA_DE_Heatmap.pdf",
  extraCellProperty = NULL,
  SeuratColorScheme = TRUE
)
}
\arguments{
\item{rca.obj}{data matrix (genes x cells)}

\item{scale}{TRUE if genes should be z-transformed before plotting (default FALSE).}

\item{width}{width of plot in inches (default is 20).}

\item{height}{height of plot in inches (default is 20).}

\item{filename}{file name of saved heatmap (default RCA_DE_Heatmap)}

\item{extraCellProperty}{vector indicating cell property to be plotted in heatmap (default NULL)}

\item{SeuratColorScheme}{Use the color scheme known from Seurat (default TRUE). Otherwise use a blue to red color scheme.}
}
\description{
Plot heatmap of DE genes
}
\examples{
\dontrun{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
RCA.pbmcs <- dataClust(RCA.pbmcs)
RCA.pbmcs <- dataDE(RCA.pbmcs)
plotDEHeatmap(RCA.pbmcs, scale = TRUE)
}
}
