% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAClusterComposition.R
\name{plotRCAClusterComposition}
\alias{plotRCAClusterComposition}
\title{Plot bar plots showing the composition of RCA clusters}
\usage{
plotRCAClusterComposition(
  rca.obj,
  deepSplit = 1,
  filename = "Cluster_Composition.pdf"
)
}
\arguments{
\item{rca.obj}{data matrix (genes x cells)}

\item{deepSplit}{provides the index of the clustering to use for different cuts in the hierarchical clustering (default 1)}

\item{filename}{file name of saved heatmap}
}
\value{
figure showing absolute and relative cluster compositions.
}
\description{
Plot bar plots showing the composition of RCA clusters
}
\examples{
\dontrun{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
RCA.pbmcs <- dataClust(RCA.pbmcs)
plotRCAClusterComposition(RCA.pbmcs,
RCA.pbmcs$clustering.out$dynamicColorsList[[1]])
}

}
