% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAHeatmap.R
\name{plotRCAHeatmap}
\alias{plotRCAHeatmap}
\title{Plot heatmap of projection to the RCA panel}
\usage{
plotRCAHeatmap(
  rca.obj,
  var.thrs = 0.1,
  width = 20,
  height = 20,
  filename = "RCA_Heatmap.pdf",
  extraCellProperty = NULL,
  SeuratColorScheme = FALSE
)
}
\arguments{
\item{rca.obj}{data matrix (genes x cells)}

\item{var.thrs}{Minimum threshold of variance for cell type correlations (default 0.1).}

\item{width}{width of plot in inches (default 20).}

\item{height}{height of plot in inches (default 20).}

\item{filename}{file name of saved heatmap (RCA_Heatmap.pdf)}

\item{extraCellProperty}{vector indicating cell property to be plotted in heatmap (default NULL)}

\item{SeuratColorScheme}{Use the color scheme known from Seurat (default FALSE). Otherwise, use a blue to red color scheme}
}
\description{
Plot heatmap of projection to the RCA panel
}
\examples{
\dontrun{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
RCA.pbmcs <- dataClust(RCA.pbmcs)
plotRCAHeatmap(RCA.pbmcs)
}

}
