% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureUMAP.R, R/plotRCAUMAP.R
\name{plotRCAUMAP}
\alias{plotRCAUMAP}
\title{Plot umap of projection to the RCA panel}
\usage{
plotRCAUMAP(
  rca.obj,
  cellPropertyList = NULL,
  folderpath = ".",
  filename = "RCA_UMAP.pdf",
  fontsize = 10
)

plotRCAUMAP(
  rca.obj,
  cellPropertyList = NULL,
  folderpath = ".",
  filename = "RCA_UMAP.pdf",
  fontsize = 10
)
}
\arguments{
\item{rca.obj}{RCA object}

\item{cellPropertyList}{list of cell properties to plot}

\item{folderpath}{path to save umap to}

\item{filename}{file name of saved umap}

\item{fontsize}{Size of the font used for plotting}
}
\description{
The presence of cell type estimates, relative ranks and confindence scores are detected automatically and are plotted accordingly.

The presence of cell type estimates, relative ranks and confindence scores are detected automatically and are plotted accordingly.
}
