% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAUMAP3D.R
\name{plotRCAUMAP3D}
\alias{plotRCAUMAP3D}
\title{Plot umap of projection to the RCA panel}
\usage{
plotRCAUMAP3D(rca.obj, cellPropertyList = NULL, filename = "RCA_UMAP.html")
}
\arguments{
\item{rca.obj}{RCA object}

\item{cellPropertyList}{list of cell properties to plot (default NULL)}

\item{filename}{file name of saved umap (filename RCA_UMAP.html)}
}
\description{
Plot umap of projection to the RCA panel
}
\examples{
\dontrun{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
RCA.pbmcs <- computeUMAP(RCA.pbmcs, nDIMS = 3)
plotRCAUMAP3D(RCA.pbmcs)
}
}
