% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scQ.R
\name{scQ}
\alias{scQ}
\title{scQ (single cell quantile normalization) for scRNAseq data}
\usage{
scQ(data, thr)
}
\arguments{
\item{data}{Expression data frame with genes in rows.}

\item{thr}{cells with detected genes less than \code{thr+1} would be  thrown away. Also after normalization
only \code{thr} number of genes would have non zero epression in each cell.}
}
\value{
normalized data as a data frame retaining orginal row and column names
}
\description{
scQ (single cell quantile normalization) for scRNAseq data
}
\examples{
norm_Data <- scQ(fpkm_data,1000) # scQ normalization
}

