% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferencePanel.R
\name{buildReferencePanel}
\alias{buildReferencePanel}
\title{Function to build a reference panel from your bulk RNA sequencing data}
\usage{
buildReferencePanel(
  bulk.rna.data,
  celltype.vec = colnames(bulk.rna.data),
  fc.thrs = 7.5,
  fdr.thrs = 0.01,
  gene.nomenclature = "SYMBOL",
  species = "HUMAN",
  folder.path = ".",
  filename = "my_reference_panel.rds"
)
}
\arguments{
\item{bulk.rna.data}{Bulk RNA sequencing data - ideally TPM normalized.}

\item{celltype.vec}{Vector of cell type names of each bulk RNA sequencing sample.}

\item{fc.thrs}{fold change threshold}

\item{fdr.thrs}{FDR threshold}

\item{gene.nomenclature}{Type of gene nomenclature. "ENS" for Ensembl IDs, "SYMBOL" for gene symbols. Default is "SYMBOL".}

\item{species}{"HUMAN" for human data, "MOUSE" for mouse data. Default is "HUMAN".}

\item{folder.path}{path to store reference panel}

\item{filename}{file name of reference panel (should end with .rds)}
}
\description{
Function to build a reference panel from your bulk RNA sequencing data
}
