% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUMAP.R
\name{computeUMAP}
\alias{computeUMAP}
\title{Compute a umap projection into 2 or 3 dimensions}
\usage{
computeUMAP(
  rca.obj,
  nDIMS = 2,
  nThreads = 1,
  nThreadsSGD = 1,
  approxPow = FALSE
)
}
\arguments{
\item{rca.obj}{RCA object}

\item{nDIMS}{number of UMAP dimensions (default 2)}

\item{nThreads}{Number of threads to use for UMAP computation (default 1)}

\item{nThreadsSGD}{Number of threads used for stochastic gradient descent (WARNING Results are not reproducible if this is > 1 (default))}

\item{approxPow}{approximate the power function in stochastic gradient descent (Default FALSE)}
}
\description{
Compute a umap projection into 2 or 3 dimensions
}
\examples{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
RCA.pbmcs <- dataLogNormalise(RCA.pbmcs)
RCA.pbmcs <- dataProject(RCA.pbmcs, method = "GlobalPanel_CellTypes")
RCA.pbmcs <- computeUMAP(RCA.pbmcs)
}
