% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRCAObject.R
\name{createRCAObject}
\alias{createRCAObject}
\title{Create Reference Class}
\usage{
createRCAObject(raw_data)
}
\arguments{
\item{raw_data}{A matrix of expression values}
}
\value{
RCA object.
}
\description{
Create Reference Class
}
