% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRCAObject.R
\name{createRCAObject}
\alias{createRCAObject}
\title{Create RCA object}
\usage{
createRCAObject(rawData, normData = NULL)
}
\arguments{
\item{rawData}{A matrix of expression values}

\item{normData}{A matrix of normalized expression values (default NULL)}
}
\value{
RCA object.
}
\description{
Create RCA object
}
\examples{
RCA.pbmcs <- createRCAObject(RCAv2::pbmc_small_counts)
print(RCA.pbmcs)

}
