% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRCAObjectFrom10X.R
\name{createRCAObjectFrom10X}
\alias{createRCAObjectFrom10X}
\title{Create RCA Object from 10X data}
\usage{
createRCAObjectFrom10X(dataDir, cellrangerVersion = 3)
}
\arguments{
\item{dataDir}{Directory containing 10X data}

\item{cellrangerVersion}{Version of cellranger used to generate data. Default is 3.0.}
}
\value{
RCA object.
}
\description{
Create RCA Object from 10X data
}
