% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFilter.R
\name{dataFilter}
\alias{dataFilter}
\title{Filter the dataset and remove genes and cells that could be of poor quality using user-defined thresholds}
\usage{
dataFilter(rca.obj, nGene.thresholds = c(100, NULL),
  nUMI.thresholds = c(1000, NULL), percent.mito.thresholds = c(0, 0.2),
  min.cell.exp = 10, plot = T, folderpath = ".",
  filename = "RCA_Filter.pdf")
}
\arguments{
\item{rca.obj}{RCA object.}

\item{nGene.thresholds}{numeric vector with lower and upper nGene thresholds}

\item{nUMI.thresholds}{numeric vector with lower and upper nUMI thresholds}

\item{percent.mito.thresholds}{numeric vector with lower and upper pMito thresholds}

\item{min.cell.exp}{minimum number of cells a gene must be expressed in}

\item{plot}{boolean - if True, plot data filter metrics}

\item{folderpath}{path to save data filter metric plots to}

\item{filename}{file name of saved plots}
}
\value{
RCA object.
}
\description{
Filter the dataset and remove genes and cells that could be of poor quality using user-defined thresholds
}
