% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLogNormalise.R
\name{dataLogNormalise}
\alias{dataLogNormalise}
\title{Log-transform and normalise data by sequencing depth}
\usage{
dataLogNormalise(rca.obj)
}
\arguments{
\item{rca.obj}{RCA object.}
}
\value{
RCA object.
}
\description{
Log-transform and normalise data by sequencing depth
}
