% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSClust.R
\name{dataSClust}
\alias{dataSClust}
\title{Generate cell clusters using Seurat graph based clustering.}
\usage{
dataSClust(rca.obj, res = 0.5)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{res}{Resolution parameter (between 0.0 and 1.0) to be used for clustering. Default: 0.5. Check the Seurat documentation for details.}
}
\value{
RCA object.
}
\description{
Generate cell clusters using Seurat graph based clustering.
}
