% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSNN.R
\name{dataSNN}
\alias{dataSNN}
\title{Generate cell clusters using hierarchical clustering and dynamic tree cutting.}
\usage{
dataSNN(rca.obj, k = 10, eps = 8, minPts = 5)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{k}{Number of cells to consider in the neirest neighbour graph. Default is 10.}

\item{eps}{Number of cells that have to be similar to connect 2 cells in a network. Default is 8}

\item{minPts}{Number of cells with at least eps neighbours to be considered a core point. Default is 5.}
}
\value{
RCA object.
}
\description{
Generate cell clusters using hierarchical clustering and dynamic tree cutting.
}
