% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellTypeFromProjection.R
\name{estimateCellTypeFromProjection}
\alias{estimateCellTypeFromProjection}
\title{Estimate the most likely cell type from the projection to the reference panel}
\usage{
estimateCellTypeFromProjection(rca.obj, confidence = NULL, ctRank = F,
  cSCompute = F)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{confidence:}{a parameter indicating the difference between z-scores. If the difference is below this threshold, the cell type will be set to unknown. Default is NULL.}

\item{ctRank:}{parameter indicating whether a relative rank coloring for each cell type shall be computed. Default is FALSE.}

\item{cSCompute:}{parameter indicating wheter the confidence score should be computed for each cell. Default is FALSE.}
}
\value{
RCA object.
}
\description{
Estimate the most likely cell type from the projection to the reference panel
}
