% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellTypeFromProjectionPerCluster.R
\name{estimateCellTypeFromProjectionPerCluster}
\alias{estimateCellTypeFromProjectionPerCluster}
\title{Estimate the most likely cell type from the projection to the reference panel}
\usage{
estimateCellTypeFromProjectionPerCluster(rca.obj, homogeneity = NULL)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{homogeneity}{a parameter indicating the homogeneity of the cluster. If the difference is below this threshold, the cell type will be set to unknown. Default is NULL.}
}
\value{
RCA object.
}
\description{
Estimate the most likely cell type from the projection to the reference panel
}
