% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterSpaceSeurat.R
\name{parameterSpaceSeurat}
\alias{parameterSpaceSeurat}
\title{Generates a data frame showcasing the dependence between resolution and cluster quantity}
\usage{
parameterSpaceSeurat(
  rca.obj,
  stepsize = 0.1,
  folderpath = ".",
  filename = "Seurat_Parameter_Space.pdf"
)
}
\arguments{
\item{rca.obj}{RCA object.}

\item{stepsize}{Stepsize used to generate an overview on clustering results.}

\item{folderpath}{Path to store the html file. Defaults to working directory.}

\item{filename}{name of the pdf file produced that visuzalizes the parameter space.}
}
\value{
a data frame holding parameter values and resulting cluster numbers.
}
\description{
Generates a data frame showcasing the dependence between resolution and cluster quantity
}
