% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterSpaceSNN.R
\name{parameterSpaceSNN}
\alias{parameterSpaceSNN}
\title{Generate cell clusters using hierarchical clustering and dynamic tree cutting.}
\usage{
parameterSpaceSNN(rca.obj, kL = c(30:50), epsL = c(5:20),
  minPtsL = c(5:10), folderpath = ".",
  filename = "Graph_based_Clustering_Parameter_Space.html")
}
\arguments{
\item{rca.obj}{RCA object.}

\item{kL}{vector of k values to be used. Default is 10-50.}

\item{epsL}{vector of values indicating the number of neighbours between two nodes such that cells are shared. Default is 5 to 20.}

\item{minPtsL}{Vector of minimum points with at least eps neighbours to be considered a core point. Default is 5 to 10.}

\item{folderpath}{Path to store the html file. Defaults to working directory.}

\item{filename}{name of the html file produced that visuzalizes the parameter space.}
}
\value{
a data frame holding parameter values and resulting cluster numbers.
}
\description{
Generate cell clusters using hierarchical clustering and dynamic tree cutting.
}
