% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterQuality.R
\name{plotClusterQuality}
\alias{plotClusterQuality}
\title{Perform cluster specific QC}
\usage{
plotClusterQuality(
  rca.obj,
  cluster.labels,
  width = 20,
  height = 20,
  folderpath = ".",
  filename = "RCA_Cluster_Quality.pdf"
)
}
\arguments{
\item{rca.obj}{RCA object}

\item{cluster.labels}{vector of cluster labels}

\item{width}{width of plot in inches. Default is 20.}

\item{height}{height of plot in inches. Default is 20.}

\item{folderpath}{path to save heatmap to}

\item{filename}{file name of saved heatmap}
}
\description{
Perform cluster specific QC
}
