% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEHeatmap.R
\name{plotDEHeatmap}
\alias{plotDEHeatmap}
\title{Plot heatmap of DE genes}
\usage{
plotDEHeatmap(
  rca.obj,
  scale = FALSE,
  width = 20,
  height = 20,
  folderpath = ".",
  filename = "RCA_DE_Heatmap.pdf",
  extraCellProperty = NULL
)
}
\arguments{
\item{rca.obj}{data matrix (genes x cells)}

\item{scale}{TRUE if genes should be z-transformed before plotting. Default is FALSE.}

\item{width}{width of plot in inches. Default is 20.}

\item{height}{height of plot in inches. Default is 20.}

\item{folderpath}{path to save heatmap to}

\item{filename}{file name of saved heatmap}

\item{extraCellProperty}{vector indicating cell property to be plotted in heatmap}
}
\description{
Plot heatmap of DE genes
}
