% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAClusterComposition.R
\name{plotRCAClusterComposition}
\alias{plotRCAClusterComposition}
\title{Plot bar plots showing the composition of RCA clusters}
\usage{
plotRCAClusterComposition(
  rca.obj,
  deepSplit = 1,
  folderpath = ".",
  filename = "Cluster_Composition.pdf"
)
}
\arguments{
\item{rca.obj}{data matrix (genes x cells)}

\item{deepSplit}{provides the index of the clustering to use for different cuts in the hierarchical clustering (default is 1)}

\item{folderpath}{path to save heatmap to}

\item{filename}{file name of saved heatmap}
}
\description{
Plot bar plots showing the composition of RCA clusters
}
