% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAClusterComposition.R
\name{plotRCAClusterComposition}
\alias{plotRCAClusterComposition}
\title{Plot bar plots showing the composition of RCA clusters}
\usage{
plotRCAClusterComposition(rca.obj, folderpath = ".",
  filename = "RCA_Heatmap.pdf")
}
\arguments{
\item{rca.obj}{data matrix (genes x cells)}

\item{folderpath}{path to save heatmap to}

\item{filename}{file name of saved heatmap}
}
\description{
Plot bar plots showing the composition of RCA clusters
}
