% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAHeatmap.R
\name{plotRCAHeatmap}
\alias{plotRCAHeatmap}
\title{Plot heatmap of projection to the RCA panel}
\usage{
plotRCAHeatmap(rca.obj, cellPropertyList = NULL, folderpath = ".",
  filename = "RCA_Heatmap.pdf")
}
\arguments{
\item{rca.obj}{data matrix (genes x cells)}

\item{cellPropertyList}{list of cell properties to plot}

\item{folderpath}{path to save heatmap to}

\item{filename}{file name of saved heatmap}
}
\description{
Plot heatmap of projection to the RCA panel
}
