% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAHeatmap.R
\name{plotRCAHeatmap}
\alias{plotRCAHeatmap}
\title{Plot heatmap of projection to the RCA panel}
\usage{
plotRCAHeatmap(
  rca.obj,
  var.thrs = 0.1,
  width = 20,
  height = 20,
  folderpath = ".",
  filename = "RCA_Heatmap.pdf",
  extraCellProperty = NULL,
  SeuratColorScheme = FALSE
)
}
\arguments{
\item{rca.obj}{data matrix (genes x cells)}

\item{var.thrs}{Minimum threshold of variance for cell type correlations.}

\item{width}{width of plot in inches. Default is 20.}

\item{height}{height of plot in inches. Default is 20.}

\item{folderpath}{path to save heatmap to}

\item{filename}{file name of saved heatmap}

\item{extraCellProperty}{vector indicating cell property to be plotted in heatmap}

\item{SeuratColorScheme}{Use the color scheme known from Seurat (default is FALSE). Otherwise, use a blue to red color scheme}
}
\description{
Plot heatmap of projection to the RCA panel
}
