% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAUMAP.R
\name{plotRCAUMAP}
\alias{plotRCAUMAP}
\title{Plot heatmap of projection to the RCA panel}
\usage{
plotRCAUMAP(rca.obj, cellPropertyList = NULL, folderpath = ".",
  filename = "RCA_UMAP.pdf")
}
\arguments{
\item{rca.obj}{RCA object}

\item{cellPropertyList}{list of cell properties to plot}

\item{folderpath}{path to save heatmap to}

\item{filename}{file name of saved heatmap}
}
\description{
Plot heatmap of projection to the RCA panel
}
