% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCAUMAP3D.R
\name{plotRCAUMAP3D}
\alias{plotRCAUMAP3D}
\title{Plot umap of projection to the RCA panel}
\usage{
plotRCAUMAP3D(rca.obj, cellPropertyList = NULL, folderpath = ".",
  filename = "RCA_UMAP.html")
}
\arguments{
\item{rca.obj}{RCA object}

\item{cellPropertyList}{list of cell properties to plot}

\item{folderpath}{path to save umap to}

\item{filename}{file name of saved umap}
}
\description{
Plot umap of projection to the RCA panel
}
