\name{RCA-package}
\alias{RCA-package}
\alias{RCA}
\docType{package}
\title{
RCA: Clustering Single Cell RNAseq Data using Reference Component Analysis
}
\description{
This package provides the Reference Component Analysis (RCA) algorithm for clustering single cell RNAseq data. The current version can be applied on data from human samples. An extention for mouse data clustering will be available soon.  
}
\details{
\tabular{ll}{
Package: \tab RCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-20\cr
License: \tab MIT\cr
}
}
\author{
Huipeng Li

Huipeng Li <lihp@gis.a-star.edu.sg> or <lihuipengsmacsb@gmail.com>
}
\references{
http://dx.doi.org/ 10.1038/ng.3818
}

\keyword{ package }