% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelComp.R
\name{RelComp}
\alias{RelComp}
\title{Reliable Component Analysis of Cliff & Caruso 1998}
\usage{
RelComp(
  data,
  varlist,
  reliab,
  retain = 5,
  criterion = 0.7,
  method = "pearson",
  use = "pairwise.complete.obs",
  file = "RelCompOutput.txt",
  digits = 3
)
}
\arguments{
\item{data}{data frame with variables in \code{varlist}.}

\item{varlist}{Character list with the variable names in \code{data} that
should be included in the analysis.}

\item{reliab}{Numeric vector of reliabilities in (0,1)}

\item{retain}{Integer. The number of components to retain.}

\item{criterion}{Value in (0,1), sets the minimum reliability}

\item{method}{See \code{\link{cor}}.}

\item{use}{See \code{\link{cor}}.}

\item{file}{Character. File name with an'.txt' extention for saving output.}
}
\description{
R implementation of Reliable Component Analysis of Cliff & Caruso 1998
}
\examples{

# Replicate the analyses of Cliff and Caruso (1998)

library(RCA)

# simulate data using the Cliff and Caruso (1998) correlation matrix
data(corr)
data <- corrSim(corr)
reliab <- c(.90, .89, .96, .86, .84, .82, .85, .76, .89, .68, .86)

# view the RelComp documentation
?RelComp

# run RelComp on the simulated data
output <- RelComp(data, names(data), reliab)

# view the reliabilities, compare to the first row in Table 8 of 
# Cliff and Caruso (1998)
round(output$reliabilities[1:5], 3)

}
